add: 컴포넨트 각각 파일에 저장
